Verify solutions in levels loaded from GitHub.
If there is a problem with any of the solutions then the level fails to
load and a stacktrace is printed to the command line. This is not really
the desired behaviour - it would be better to display a dialog that asked
if you wanted to play it anyway.

Also outstanding is finding out if any solution actually wins the level
and displaying that information to the user.